+++ src/GRA.Domain.Model/User.cs
@@ -65,9 +65,6 @@ namespace GRA.Domain.Model
         public int? Age { get; set; }
         [DisplayName("School")]
         public int? SchoolId { get; set; }
-        public int? EnteredSchoolId { get; set; }
-        [DisplayName("School Name")]
-        public string EnteredSchoolName { get; set; }
 
         public DateTime? AchievedAt { get; set; }
         public DateTime? LastBroadcast { get; set; }